6494 Move CurrencyCode to Yosemite
This is to allow the use of CurrencyCode in CardPresentPaymentsConfiguration
