+++ WooCommerce/Classes/Tools/Currency/CurrencySettings.swift
@@ -7,63 +7,6 @@ public class CurrencySettings {
 
     // MARK: - Enums
 
-    /// The 3-letter country code for supported currencies
-    ///
-    public enum CurrencyCode: String, CaseIterable {
-        // A
-        case AED, AFN, ALL, AMD, ANG, AOA, ARS, AUD, AWG, AZN,
-        // B
-        BAM, BBD, BDT, BGN, BHD, BIF, BMD, BND, BOB, BRL, BSD, BTC, BTN, BWP, BYR, BYN, BZD,
-        // C
-        CAD, CDF, CHF, CLP, CNY, COP, CRC, CUC, CUP, CVE, CZK,
-        // D
-        DJF, DKK, DOP, DZD,
-        // E
-        EGP, ERN, ETB, EUR, FJD,
-        // F
-        FKP,
-        // G
-        GBP, GEL, GGP, GHS, GIP, GMD, GNF, GTQ, GYD,
-        // H
-        HKD, HNL, HRK, HTG, HUF,
-        // I
-        IDR, ILS, IMP, INR, IQD, IRR, IRT, ISK,
-        // J
-        JEP, JMD, JOD, JPY,
-        // K
-        KES, KGS, KHR, KMF, KPW, KRW, KWD, KYD, KZT,
-        // L
-        LAK, LBP, LKR, LRD, LSL, LYD,
-        // M
-        MAD, MDL, MGA, MKD, MMK, MNT, MOP, MRO, MUR, MVR, MWK, MXN, MYR, MZN,
-        // N
-        NAD, NGN, NIO, NOK, NPR, NZD,
-        // O
-        OMR,
-        // P
-        PAB, PEN, PGK, PHP, PKR, PLN, PRB, PYG,
-        // Q
-        QAR,
-        // R
-        RMB, RON, RSD, RUB, RWF,
-        // S
-        SAR, SBD, SCR, SDG, SEK, SGD, SHP, SLL, SOS, SRD, SSP, STD, SYP, SZL,
-        // T
-        THB, TJS, TMT, TND, TOP, TRY, TTD, TWD, TZS,
-        // U
-        UAH, UGX, USD, UYU, UZS,
-        // V
-        VEF, VND, VUV,
-        // W
-        WST,
-        // X
-        XAF, XCD, XOF, XPF,
-        // Y
-        YER,
-        // Z
-        ZAR, ZMW
-    }
-
     /// Designates where the currency symbol is located on a formatted price
     ///
     public enum CurrencyPosition: String {
