+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/Create Shipping Label Form/Customs/ShippingLabelCustomsFormListViewModel.swift
@@ -64,7 +64,7 @@ final class ShippingLabelCustomsFormListViewModel: ObservableObject {
         self.allCountries = countries
         self.destinationCountry = destinationCountry
         let currencySymbol: String = {
-            guard let currencyCode = CurrencySettings.CurrencyCode(rawValue: order.currency) else {
+            guard let currencyCode = CurrencyCode(rawValue: order.currency) else {
                 return ""
             }
             return ServiceLocator.currencySettings.symbol(from: currencyCode)
